refactor: move commands tests to a different folder
